Fix numpy dep warning

+++ topo/spectral/umap_layouts.py
@@ -694,7 +694,7 @@ def _optimize_layout_aligned_euclidean_single_epoch(
         if e_p_s.shape[0] >= max_n_edges:
             max_n_edges = e_p_s.shape[0]
 
-    embedding_order = np.arange(n_embeddings).astype(np.int32)
+    embedding_order = np.arange(n_embeddings).astype(int)
     np.random.shuffle(embedding_order)
 
     for i in range(max_n_edges):
